Print a Path Summary for a Fit from sgdnet

Print a Path Summary for a Fit from sgdnet

# S3 method for sgdnet
print(x, ...)

Arguments

x
an object of class 'sgdnet', typically the result of calling sgdnet()

...
other arguments passed to print()

Value

Prints (and return invisibly) a table of the regularization path from the fit with the following columns:

Df	pseudo-degrees of freedom, namely the number of non-zero coefficients

%Dev	the percentage of deviance explained

Lambda	the \(\lambda\) value (regularization strength) of the fit

See also

Examples

fit <- sgdnet(with(mtcars, cbind(drat, hp)), mtcars$disp)
print(fit, digits = 1)

#> 
#> Call:  sgdnet.default(x = with(mtcars, cbind(drat, hp)), y = mtcars$disp) 
#> 
#>     Df   %Dev Lambda
#> s0   0 -7e-04  96.49
#> s1   1  1e-01  87.91
#> s2   1  2e-01  80.10
#> s3   2  3e-01  72.99
#> s4   2  4e-01  66.50
#> s5   2  4e-01  60.60
#> s6   2  5e-01  55.21
#> s7   2  5e-01  50.31
#> s8   2  6e-01  45.84
#> s9   2  6e-01  41.77
#> s10  2  6e-01  38.06
#> s11  2  7e-01  34.68
#> s12  2  7e-01  31.59
#> s13  2  7e-01  28.79
#> s14  2  7e-01  26.23
#> s15  2  7e-01  23.90
#> s16  2  7e-01  21.78
#> s17  2  7e-01  19.84
#> s18  2  8e-01  18.08
#> s19  2  8e-01  16.47
#> s20  2  8e-01  15.01
#> s21  2  8e-01  13.68
#> s22  2  8e-01  12.46
#> s23  2  8e-01  11.35
#> s24  2  8e-01  10.35
#> s25  2  8e-01   9.43
#> s26  2  8e-01   8.59
#> s27  2  8e-01   7.83
#> s28  2  8e-01   7.13
#> s29  2  8e-01   6.50
#> s30  2  8e-01   5.92
#> s31  2  8e-01   5.39
#> s32  2  8e-01   4.92
#> s33  2  8e-01   4.48
#> s34  2  8e-01   4.08
#> s35  2  8e-01   3.72
#> s36  2  8e-01   3.39
#> s37  2  8e-01   3.09
#> s38  2  8e-01   2.81
#> s39  2  8e-01   2.56
#> s40  2  8e-01   2.34
#> s41  2  8e-01   2.13
#> s42  2  8e-01   1.94
#> s43  2  8e-01   1.77
#> s44  2  8e-01   1.61
#> s45  2  8e-01   1.47
#> s46  2  8e-01   1.34
#> s47  2  8e-01   1.22
#> s48  2  8e-01   1.11
#> s49  2  8e-01   1.01
#> s50  2  8e-01   0.92
#> s51  2  8e-01   0.84
#> s52  2  8e-01   0.76
#> s53  2  8e-01   0.70
#> s54  2  8e-01   0.63
#> s55  2  8e-01   0.58
#> s56  2  8e-01   0.53
#> s57  2  8e-01   0.48
#> s58  2  8e-01   0.44
#> s59  2  8e-01   0.40
#> s60  2  8e-01   0.36
#> s61  2  8e-01   0.33
#> s62  2  8e-01   0.30
#> s63  2  8e-01   0.27
#> s64  2  8e-01   0.25
#> s65  2  8e-01   0.23
#> s66  2  8e-01   0.21
#> s67  2  8e-01   0.19
#> s68  2  8e-01   0.17
#> s69  2  8e-01   0.16
#> s70  2  8e-01   0.14
#> s71  2  8e-01   0.13
#> s72  2  8e-01   0.12
#> s73  2  8e-01   0.11
#> s74  2  8e-01   0.10
#> s75  2  8e-01   0.09
#> s76  2  8e-01   0.08
#> s77  2  8e-01   0.07
#> s78  2  8e-01   0.07
#> s79  2  8e-01   0.06
#> s80  2  8e-01   0.06
#> s81  2  8e-01   0.05
#> s82  2  8e-01   0.05
#> s83  2  8e-01   0.04
#> s84  2  8e-01   0.04
#> s85  2  8e-01   0.04
#> s86  2  8e-01   0.03
#> s87  2  8e-01   0.03
#> s88  2  8e-01   0.03
#> s89  2  8e-01   0.02
#> s90  2  8e-01   0.02
#> s91  2  8e-01   0.02
#> s92  2  8e-01   0.02
#> s93  2  8e-01   0.02
#> s94  2  8e-01   0.02
#> s95  2  8e-01   0.01
#> s96  2  8e-01   0.01
#> s97  2  8e-01   0.01
#> s98  2  8e-01   0.01
#> s99  2  8e-01   0.01